Add core module with audio recording and transcription
Introduced the 'core' package containing modules for configuration management, audio recording, caching, FastWhisper settings dialog, and transcription threading. Updated .gitignore to exclude the /.claude file. Modified the main application logic to use the new core modules, centralizing audio capture, processing, and transcription features for the VoxSynopsis application.
