+++ .gitignore
@@ -420,3 +420,4 @@ FodyWeavers.xsd
 gravacoes/
 Tunning/
 config.json
+/.claude
